moved CloseVaultCard to vault compoments
