+++ components/vault/CloseVaultCard.tsx
